--- src/ragas/metrics/domain_specific_rubrics/__init__.py
@@ -1,15 +1,0 @@
-from ragas.metrics.domain_specific_rubrics.with_reference import (
-    RubricsScoreWithReference,
-    rubrics_score_with_reference,
-)
-from ragas.metrics.domain_specific_rubrics.without_reference import (
-    RubricsScoreWithoutReference,
-    rubrics_score_without_reference,
-)
-
-__all__ = [
-    "RubricsScoreWithReference",
-    "RubricsScoreWithoutReference",
-    "rubrics_score_with_reference",
-    "rubrics_score_without_reference",
-]
